separate log from class
