+++ spark_api/src/spark_proxy.py
@@ -1,3 +1,5 @@
+from collections.abc import Generator
+
 import sparkapi_pb2
 from sparkapi_pb2_grpc import SparkApiServicer
 
@@ -6,7 +8,7 @@ from PipelinePlan import SessionPlanner
 from spark_session_init import clientSessionTable, spark
 
 
-class SparkApiServicer(SparkApiServicer):
+class SparkRpcApi(SparkApiServicer):
     def createSession(self, req: sparkapi_pb2.NewSessionRequest, unused_context) -> sparkapi_pb2.NewSessionResponse:
         clientSessionTable.add(req.id, ClientSession(req.id, req.name))
 
@@ -19,7 +21,9 @@ class SparkApiServicer(SparkApiServicer):
     def submit_LoadDatasetNode(self, req: sparkapi_pb2.LoadDatasetNodeRequest, unused_context) -> sparkapi_pb2.SparkTransformResponse:
         session = clientSessionTable.get_session(req.session_id)
 
-        node = session.submit_node(node_class='LoadNode', session_id=session.id, input_metadata=getattr(req, req.WhichOneof('input_metadata')))
+        node = session.submit_node(
+            node_class='LoadNode', session_id=session.id, input_metadata=getattr(req, req.WhichOneof('input_metadata'))
+        )
 
         session.plan = SessionPlanner(req.session_id, node)
         return sparkapi_pb2.SparkTransformResponse(
@@ -28,7 +32,9 @@ class SparkApiServicer(SparkApiServicer):
             columns=node.columns,
         )
 
-    def submit_LoadFromSessionNode(self, req: sparkapi_pb2.LoadFromSessionNodeRequest, unused_context) -> sparkapi_pb2.SparkTransformResponse:
+    def submit_LoadFromSessionNode(
+        self, req: sparkapi_pb2.LoadFromSessionNodeRequest, unused_context
+    ) -> sparkapi_pb2.SparkTransformResponse:
         session = clientSessionTable.get_session(req.session_id)
 
         parent_session = clientSessionTable.get_session(req.input_session_id)
@@ -176,7 +182,7 @@ class SparkApiServicer(SparkApiServicer):
             schema=summary['schema'],
         )
 
-    def previewDataset(self, req: sparkapi_pb2.PreviewDatasetRequest, unused_context) -> sparkapi_pb2.DatasetResponse:
+    def previewDataset(self, req: sparkapi_pb2.PreviewDatasetRequest, unused_context) -> Generator[sparkapi_pb2.DatasetResponse]:
         session = clientSessionTable.get_session(req.session_id)
         node = session.plan.get_node_by_id(req.node_id)
